requirements.txt
astroid==2.11.4
autopep8==1.6.0
colorama==0.4.4
dill==0.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.13.8
toml==0.10.2
tomli==2.0.1
wrapt==1.14.1
